+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add env_logger to support controlling log output through the RUST_LOG environment variable.
 - Proxy requests to the remote server when running a remote HTML step. We use this to load static assets.
 - Add version number to the JSON config that steps receive.
+- Prefix HTML step routes with /srvc. The old routes are deprecated, but will still work for now.
 
 ## v0.13.0 (2023-01-24)